+++ composer.json
@@ -20,15 +20,12 @@
   "require": {
     "php": ">=5.5",
     "doctrine/orm": "~2.5",
-    "doctrine/doctrine-module": "~0.8",
+    "doctrine/doctrine-module": "~0.9",
     "zfr/zfr-stripe-module": "~3.0",
-    "ddeboer/vatin": "~1.3.0",
     "zendframework/zend-mvc": "~2.2",
-    "zendframework/zend-http": "~2.2",
     "zendframework/zend-eventmanager": "~2.2",
     "zendframework/zend-modulemanager": "~2.2",
-    "zendframework/zend-servicemanager": "~2.2",
-    "zendframework/zend-validator": "~2.2"
+    "zendframework/zend-servicemanager": "~2.2"
   },
   "require-dev": {
     "phpunit/phpunit": "~4.1"
