--- config/doctrine/ZfrCash.Entity.Invoice.dcm.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<doctrine-mapping
-        xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
-    http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-
-    <entity name="ZfrCash\Entity\Invoice" table="cash_invoices">
-        <id name="id" type="integer">
-            <generator strategy="AUTO" />
-        </id>
-
-        <field name="stripeId" column="stripe_id" type="string" length="255" unique="true" />
-        <many-to-one field="payer" target-entity="ZfrCash\Entity\CustomerInterface" fetch="EAGER">
-            <join-column on-delete="CASCADE" />
-        </many-to-one>
-        <field name="periodStart" column="period_start" type="datetime" />
-        <field name="periodEnd" column="period_end" type="datetime" />
-        <field name="startingBalance" column="starting_balance" type="integer" />
-        <field name="endingBalance" column="ending_balance" type="integer" />
-        <field name="subtotal" type="integer" />
-        <field name="total" type="integer" />
-        <field name="applicationFee" column="application_fee" type="integer" />
-        <field name="tax" type="integer" />
-        <field name="taxPercent" column="tax_percent" type="decimal" precision="5" scale="2" />
-        <field name="amountDue" column="amount_due" type="integer" />
-        <field name="currency" type="string" length="3" />
-        <field name="closed" type="boolean" />
-        <field name="paid" type="boolean" />
-        <field name="forgiven" type="boolean" />
-        <field name="attemptCount" column="attempt_count" type="smallint" />
-        <field name="description" type="string" length="1000" nullable="true" />
-        <field name="vatNumber" column="vat_number" type="string" length="60" nullable="true" />
-        <field name="vatCountry" column="vat_country" type="string" length="2" nullable="true" />
-        <field name="exportUrl" column="export_url" type="string" nullable="true" />
-        <field name="createdAt" column="created_at" type="datetime" />
-    </entity>
-</doctrine-mapping>
