+++ config/doctrine/ZfrCash.Entity.Plan.dcm.xml
@@ -22,7 +22,7 @@
         <field name="intervalCount" column="interval_count" type="smallint" />
         <field name="trialPeriodDays" column="trial_period_days" type="smallint" nullable="true" />
         <field name="features" type="json_array" />
-        <one-to-many field="limits" target-entity="ZfrCash\Entity\PlanLimit" fetch="EAGER" mapped-by="plan">
+        <one-to-many field="metadata" target-entity="ZfrCash\Entity\PlanMetadata" fetch="EAGER" mapped-by="plan">
             <cascade>
                 <cascade-persist />
             </cascade>
