+++ config/doctrine/ZfrCash.Entity.PlanMetadata.dcm.xml
@@ -5,12 +5,12 @@
         xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
     http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-    <entity name="ZfrCash\Entity\PlanLimit" table="cash_plan_limits">
+    <entity name="ZfrCash\Entity\PlanMetadata" table="cash_plan_metadata">
         <id name="id" type="integer">
             <generator strategy="AUTO" />
         </id>
 
-        <many-to-one field="plan" target-entity="ZfrCash\Entity\Plan" inversed-by="limits">
+        <many-to-one field="plan" target-entity="ZfrCash\Entity\Plan" inversed-by="metadata">
             <join-column on-delete="CASCADE" />
         </many-to-one>
         <field name="key" type="string" />
