--- src/Factory/InvoiceServiceFactory.php
@@ -1,58 +1,0 @@
-<?php
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license.
- */
-
-namespace ZfrCash\Factory;
-
-use Zend\ServiceManager\FactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
-use ZfrCash\Entity\CustomerInterface;
-use ZfrCash\Entity\Invoice;
-use ZfrCash\Entity\Subscription;
-use ZfrCash\Options\ModuleOptions;
-use ZfrCash\Service\InvoiceService;
-use ZfrStripe\Client\StripeClient;
-
-/**
- * @author  Michaël Gallego <mic.gallego@gmail.com>
- * @licence MIT
- */
-class InvoiceServiceFactory implements FactoryInterface
-{
-    /**
-     * {@inheritDoc}
-     */
-    public function createService(ServiceLocatorInterface $serviceLocator)
-    {
-        /** @var ModuleOptions $moduleOptions */
-        $moduleOptions = $serviceLocator->get(ModuleOptions::class);
-
-        /** @var \Doctrine\Common\Persistence\ObjectManager $objectManager */
-        $objectManager = $serviceLocator->get($moduleOptions->getObjectManager());
-
-        /** @var \Doctrine\Common\Persistence\ObjectRepository $invoiceRepository */
-        $invoiceRepository = $objectManager->getRepository(Invoice::class);
-
-        /** @var \ZfrCash\Repository\CustomerRepositoryInterface $customerRepository */
-        $customerRepository = $objectManager->getRepository(CustomerInterface::class);
-
-        /** @var StripeClient $stripeClient */
-        $stripeClient = $serviceLocator->get(StripeClient::class);
-
-        return new InvoiceService($objectManager, $invoiceRepository, $customerRepository, $stripeClient);
-    }
-}
